requirements.txt
streamlit
pymupdf
google-generativeai
python-dotenv
pydantic
